docs: 패키지 분리 및 정렬
