+++ src/main/java/com/RoutineGongJakSo/BE/social/Dto/KakaoUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.RoutineGongJakSo.BE.social.socialDto;
+package com.RoutineGongJakSo.BE.social.Dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
